Ai-multi-agent - learning -assistant/requirements.txt
streamlit
google-generativeai
scikit-learn
numpy
